% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_compatibility}
\alias{check_compatibility}
\title{Check compatibility within the data}
\usage{
check_compatibility(count_mat, meta_df)
}
\arguments{
\item{count_mat}{A count matrix}

\item{meta_df}{A data frame generated from the sample-sheet.}
}
\value{
compatibility a logical value TRUE for compatible FALSE otherwise.
}
\description{
Check compatibility within the data
}
\examples{
\dontrun{
count_mat <- read_count_mat("gene-count-matrix.csv")
meta_df <- read_meta_data("sample-sheet.txt")
check_compatibility(count_mat, meta_df)
}
}
