% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_qmd.R
\name{compose_qmd}
\alias{compose_qmd}
\title{Compose the Quarto markdown file.}
\usage{
compose_qmd(ofn = NULL, contrasts_df, thresh = 0.05, label_n = 30)
}
\arguments{
\item{ofn}{Name of the output file.  A temporary file will be generated}
}
\value{
ofn Path to the output file.
}
\description{
Compose the Quarto markdown file.
}
\examples{
\dontrun{
ofn <- compose_qmd()
}
}
