% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{extract_comparison}
\alias{extract_comparison}
\title{Extract compairson from gven result file.}
\usage{
extract_comparison(res_df)
}
\arguments{
\item{res_df}{result data frame}
}
\value{
str comparison
}
\description{
Extract compairson from gven result file.
}
