% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq-related-functions.R
\name{get_all_results}
\alias{get_all_results}
\title{Get results for all contrasts.}
\usage{
get_all_results(dds)
}
\arguments{
\item{dds}{A DESeqExperiment object.}
}
\value{
all_res_df A data-frame with results for all the contrasts.
}
\description{
Get results for all contrasts.
}
\examples{
\dontrun{
all_res_df <- get_all_results(dds)
}
}
