% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq-related-functions.R
\name{get_comp_from_dds}
\alias{get_comp_from_dds}
\title{Extract possible comparisons from dds colData slot}
\usage{
get_comp_from_dds(dds)
}
\arguments{
\item{dds}{A DESeqDataSet Object}
}
\value{
comb_df a data frame with possible combination of comparisons
}
\description{
Extract possible comparisons from dds colData slot
}
\examples{
\dontrun{
comb_df <- get_comp_from_dds(dds)
}
}
