% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Get contrasts for analysis}
\usage{
get_contrasts(fn = NULL, meta_df = NULL)
}
\arguments{
\item{fn}{name of the contrasts file.}

\item{meta_df}{meta-data data-frame we use this if contrasts file is not provided.}
}
\value{
contrasts_df a contarsts data-frame that will be used to extract data from the DESeq2 object.
}
\description{
Get contrasts for analysis
}
\examples{
\dontrun{
contrasts_df <- get_contrasts("contrasts.txt", meta_df)
contarsts_df <- get_contrasts(meta_df = meta_df)
}
}
