% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq-related-functions.R
\name{get_results_from_dds}
\alias{get_results_from_dds}
\title{Get results from a dds object}
\usage{
get_results_from_dds(dds = dds, df = NULL)
}
\arguments{
\item{dds}{An object of class DESeqExperiment}

\item{df}{A data frame with contrasts, one on each row. Each row contains only two columns, named c1, and c2}
}
\value{
Gives a data-frame for the results with those comparisons.
}
\description{
Get results from a dds object
}
\examples{
\dontrun{
df <- get_results_from_dds(dds, condition_df)
}
}
