% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_make_all_plots.R
\name{make_all_plots}
\alias{make_all_plots}
\title{Title}
\usage{
make_all_plots(res_df, vs_data, thresh = 0.05, label_n = 30, op_dir = NULL)
}
\arguments{
\item{res_df}{A result data frame for a single comparison}

\item{vs_data}{A vst stabilized data object.}

\item{thresh}{Threshold to be used for calling significant genes}

\item{label_n}{Number of genes to be lablled, or number of boxes in boxplot}

\item{op_dir}{Output directory where the plots should be saved.}
}
\value{
plot_l A list of plots.
}
\description{
Title
}
\examples{
\dontrun{
plot_l <- make_all_plots(res_df, vs_data, thresh, op_dir)
}
}
