% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_analysis.R
\name{make_analysis}
\alias{make_analysis}
\title{Create the analysis part of the text}
\usage{
make_analysis()
}
\value{
analysis_str analysis string that will put in place all the correct bits and pieces to actually perform the analysis.
}
\description{
Create the analysis part of the text
}
\examples{
\dontrun{
analysis_str <- make_analysis()
}
}
