% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_box_multi}
\alias{make_box_multi}
\title{Create multipanel boxplot one gene per panel}
\usage{
make_box_multi(res_df, ...)
}
\arguments{
\item{res_df}{Single result data frame}

\item{...}{Other parameters to make_box_plot_df function}
}
\value{
multi-faceted-plot a multifacted boxplot ggplot object
}
\description{
Create multipanel boxplot one gene per panel
}
\examples{
\dontrun{
make_box_multi <- make_box_multi(res_df)
}
}
