% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_box_plot_df}
\alias{make_box_plot_df}
\title{Title}
\usage{
make_box_plot_df(res_df, vs_data, meta_df, thresh = 0.05, label_n = 30)
}
\arguments{
\item{res_df}{Results data frame for one comparison}

\item{vs_data}{Variance stabilized count data for the experiment}

\item{meta_df}{The meta-data frame}

\item{thresh}{Threshold to determne significance. Default padj <= 0.05}

\item{label_n}{Number of genes to plot.}
}
\value{
l a list of information to make plots
}
\description{
Title
}
\examples{
\dontrun{
use_l <- make_box_plot_df(res_df, vs_data, meta_df)
}
}
