% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_box_single}
\alias{make_box_single}
\title{Make a single panel boxplot}
\usage{
make_box_single(res_df, ...)
}
\arguments{
\item{res_df}{Results data frame for one comparison}

\item{...}{Other parameters to get the boxplot data-frame}
}
\value{
boxplot a ggplot2 object
}
\description{
Make a single panel boxplot
}
\examples{
\dontrun{
box_p <- make_box_single(res_df, vs_data, meta_df)
}
}
