% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comaparisons.R
\name{make_comparisons}
\alias{make_comparisons}
\title{Create combinations of comparions from given condition vector}
\usage{
make_comparisons(v, n = 2)
}
\arguments{
\item{v}{A character vector with at least 2 conditions}

\item{n}{Number of conditions to use in combination, default is 2}
}
\value{
combination_df A data frame of possible combinations
}
\description{
Create combinations of comparions from given condition vector
}
\examples{
combination_df <- make_comparisons(c("Control", "Test1", "Test2"))
}
