% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq-related-functions.R
\name{make_deseq_object}
\alias{make_deseq_object}
\title{Create DESeq2 object from counts and sample-sheet}
\usage{
make_deseq_object(count_mat, meta_df)
}
\arguments{
\item{count_mat}{count matrix from the count data}

\item{meta_df}{meta-data data-frame from sample-sheet}
}
\value{
A DESeq2 object that can be used for further analysis
}
\description{
Create DESeq2 object from counts and sample-sheet
}
\examples{
\dontrun{
dds <- make_deseq_object(count_mat, meta_df)
}
}
