% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Title}
\usage{
make_heatmap(res_df, meta_df, vs_data, thresh = 0.05)
}
\arguments{
\item{res_df}{Result data frame for a single comparison}

\item{meta_df}{Meta-data used in the analysis}

\item{vs_data}{Normalized gene expression after vst transformation}

\item{thresh}{Threshold to call significantly differentially expressed genes.}
}
\value{
op_l A list of objects that can be further manipulated.
}
\description{
Title
}
\examples{
\dontrun{
op_l <- make_heatmap(res_df, meta_df, vs_data) # will save a heatmap.png in the current directory
}
}
