% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_template_parts.R
\name{make_overall_scope_bit}
\alias{make_overall_scope_bit}
\title{Make the bit for overall analysis}
\usage{
make_overall_scope_bit(contrasts_df)
}
\arguments{
\item{contrasts_df}{Data frame of contrasts.}
}
\value{
str_glue object that will be printed in the report.
}
\description{
Make the bit for overall analysis
}
\examples{
\dontrun{
  scope_bit <- make_overall_scope_bit(contrasts_df)
}
}
