% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{make_pca_plot}
\alias{make_pca_plot}
\title{Make PCA plot from a pca object}
\usage{
make_pca_plot(pca_obj, meta_df, use_title, pc_1 = 1, pc_2 = 2)
}
\arguments{
\item{pca_obj}{PCA Object}

\item{meta_df}{meta-data data frame. Can be a colData from dds}

\item{use_title}{The title for the subplot}

\item{pc_1}{First PC to be used in the scatter plot.}

\item{pc_2}{Second PC to be used in the scatter plot.}
}
\value{
p ggplot2 object
}
\description{
Make PCA plot from a pca object
}
\examples{
\dontrun{
p <- make_pca_plot(pca_obj, meta_df, use_title = "PCA Plot")
}
}
