% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{make_pca_plot_l}
\alias{make_pca_plot_l}
\title{Make a PCA plot from a DESeq2 object}
\usage{
make_pca_plot_l(dds, vs_data, contrasts_df, thresh = 0.05)
}
\arguments{
\item{dds}{A DESeqExperiment object}

\item{vs_data}{A vst transformed data object}

\item{contrasts_df}{A data frame of contrasts of interest}

\item{thresh}{Threshold to call significant genes (default 0.05)}
}
\value{
pca_plot_l a list of ggplot2 objects.
}
\description{
Make a PCA plot from a DESeq2 object
}
\examples{
\dontrun{
pca_plot_l <- make_pca_plot_l(dds, vs_data, contrasts_df)
}
}
