% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_plot_df}
\alias{make_plot_df}
\title{Get the plotting data frame.}
\usage{
make_plot_df(res_df, thresh = 0.05, label_n = 30)
}
\arguments{
\item{res_df}{Results data frame for one comparison}

\item{thresh}{Threshold to determine significance. Default 0.05 or less.}

\item{label_n}{label these many genes.}
}
\value{
l a list of data frames that can be used to make plots.
}
\description{
Get the plotting data frame.
}
\examples{
\dontrun{
use_l <- make_plot_df(res_df)
}
}
