% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{make_scree_plot}
\alias{make_scree_plot}
\title{Scree plot to explain importance of the PCs calculated}
\usage{
make_scree_plot(pca_obj)
}
\arguments{
\item{pca_obj}{An object generated with prcomp function call}
}
\value{
p ggplot2 object
}
\description{
Scree plot to explain importance of the PCs calculated
}
\examples{
\dontrun{
p <- make_scree_plot(pca_obj)
}
}
