% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_template_parts.R
\name{make_setup_qmd}
\alias{make_setup_qmd}
\title{Make setup part of the qmd file}
\usage{
make_setup_qmd()
}
\value{
setup_qmd string that holds some text to setup the analysis.
}
\description{
Make setup part of the qmd file
}
\examples{
\dontrun{
setup_qmd <- make_setup_qmd()
}
}
