% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_template_parts.R
\name{make_single_comparison_bit}
\alias{make_single_comparison_bit}
\title{Make the bits to show graphs for comparisons.}
\usage{
make_single_comparison_bit(res_df, thresh = 0.05, label_n = 30, comp_n)
}
\arguments{
\item{res_df}{Result data frame for single result.}

\item{thresh}{Adjusted p-value cutoff to call significant genes.}

\item{label_n}{Number of genes to label default is 30}

\item{comp_n}{Number of comparisons to annotated.}
}
\value{
A bit that will display graphs for one comparison at a time.
}
\description{
Make the bits to show graphs for comparisons.
}
\examples{
\dontrun{
make_single_comparison_bit(res_df)
}
}
