% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_upset_df}
\alias{make_upset_df}
\title{Make df for an upSet plot}
\usage{
make_upset_df(dds, meta_df, contrasts_df = NULL, thresh = 0.05)
}
\arguments{
\item{dds}{A DESeqExperiment object}

\item{meta_df}{A meta-data data frame.}

\item{contrasts_df}{A contrasts data frame if specific contrasts in mind.  Can be null in which case all possible combinations will be used for plotting the graph.}

\item{thresh}{a threshold for padj value (default 0.05)}
}
\value{
upset_df A data frame that can be used to make an upSet plot
}
\description{
Make df for an upSet plot
}
\examples{
\dontrun{
upset_df <- make_upset_df(dds, meta_df, contrasts_df)
}
}
