% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_upset_plot}
\alias{make_upset_plot}
\title{Create an upSet plot}
\usage{
make_upset_plot(upset_df)
}
\arguments{
\item{upset_df}{A data frame with necessary list column. Generated by make_upset_df function call.}
}
\value{
upset_plot A ggplot2 object with upset parameters.
}
\description{
Create an upSet plot
}
\examples{
\dontrun{
upset_plot <- make_upset_plot(upset_df) # when you have created the upset_df
upset_plot <- make_upset_plot(make_upset_df(dds, meta_df, contrasts_df, thresh))
# when you want to get to it directly.
}
}
