% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{make_volcano}
\alias{make_volcano}
\title{Make a volcano plot from the result data-frame.}
\usage{
make_volcano(res_df, thresh = 0.05)
}
\arguments{
\item{res_df}{Results data frame.}

\item{thresh}{Threshold to determine sigificance. This is a cutoff used on adjusted pvalue.}
}
\value{
volcano_p a volcano plot, a ggplot2 graph/object
}
\description{
Make a volcano plot from the result data-frame.
}
\examples{
\dontrun{
volcano_p <- make_volcano(res_df)
}
}
