% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{my_rnaseq_theme}
\alias{my_rnaseq_theme}
\title{My theme for plotting}
\usage{
my_rnaseq_theme()
}
\value{
themed plot
}
\description{
My theme for plotting
}
\examples{
\dontrun{
p + my_rnaseq_theme()
}
}
