% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoReport-package.R
\docType{package}
\name{quartoReport-package}
\alias{quartoReport}
\alias{quartoReport-package}
\title{quartoReport: Generate Quarto Report for RNA-Seq Analysis}
\description{
Generate interactive RNA-Seq differential analysis report using Quarto. This allows generation of multiple outputs from a single source file.
}
\author{
\strong{Maintainer}: Sameet Mehta \email{sameet.mehta@yale.edu} (\href{https://orcid.org/0000-0002-3029-3217}{ORCID})

}
\keyword{internal}
