% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count_mat.R
\name{read_count_mat}
\alias{read_count_mat}
\title{Read in the counts}
\usage{
read_count_mat(fn)
}
\arguments{
\item{fn}{}
}
\value{
A matrix of counts of integers.
}
\description{
Read in the counts
}
\examples{
count_mat <- read_count_mat("genes_count_mat.csv")
}
