% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deseq.R
\name{save_hm_png}
\alias{save_hm_png}
\title{save function for heatmap (png)}
\usage{
save_hm_png(use_l, use_dir)
}
\arguments{
\item{use_l}{Output from the make heatmap function call}

\item{use_dir}{Directory where you want to heatmap saved.}
}
\value{
returns nothing write the heatmap to the given folder with appropriate name.
}
\description{
save function for heatmap (png)
}
\examples{
\dontrun{
save_hm_pdf(use_l, "~/Desktop")
}
}
