% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.R
\name{save_objects}
\alias{save_objects}
\title{Save computed DESEq2 objects.}
\usage{
save_objects(obj, op_dir = NULL)
}
\arguments{
\item{obj}{Object to be saved, either a DESEq2 experiment or vst data}

\item{op_dir}{Directory where the files will be saved.}
}
\value{
no return value the objects are saved to the output directory
}
\description{
Save computed DESEq2 objects.
}
\examples{
\dontrun{
save_objects(dds, op_dir = ".")
}
}
