% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq-related-functions.R
\name{save_results}
\alias{save_results}
\title{Save result files.}
\usage{
save_results(res_df, op_dir, thresh = 0.05)
}
\arguments{
\item{res_df}{Result data frame for a single result}

\item{op_dir}{Output directory}

\item{thresh}{Threshold to determine significance (default padj <= 0.05)}
}
\value{
No return value.
}
\description{
Save result files.
}
\examples{
\dontrun{
save_results(res_df, op_dir = tempdir())
}
}
