% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq-related-functions.R
\name{stabilize_count_data}
\alias{stabilize_count_data}
\title{Stabilize count data using vst}
\usage{
stabilize_count_data(dds)
}
\arguments{
\item{dds}{a DESeq2 object}
}
\value{
vs_data a count object with vst transformed counts.
}
\description{
Stabilize count data using vst
}
\examples{
\dontrun{
vs_data <- stabilize_count_data(dds)
}
}
