% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meta_data.R
\name{use_meta_data}
\alias{use_meta_data}
\title{Read Meta data (sample sheet)}
\usage{
use_meta_data(fn)
}
\arguments{
\item{fn}{Tab separated values file with samples and conditions}
}
\value{
meta_df a data.frame object.
}
\description{
Read Meta data (sample sheet)
}
\examples{
\dontrun{
meta_df <- read_meta_data(file.path("tests/test_data/sample-sheet.txt"))
}
}
