% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_template_parts.R
\name{make_qc_bit}
\alias{make_qc_bit}
\title{Create the part to plot the QC file}
\usage{
make_qc_bit(fn = NULL)
}
\arguments{
\item{fn}{name of the rnaseq metrics file if available.}
}
\value{
str A string that will correctly plot the QC of the rnaseq metrics.
}
\description{
Create the part to plot the QC file
}
\examples{
\dontrun{
qc_bit <- make_qc_bit()
}
}
