% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count_mat.R
\name{read_count_mat}
\alias{read_count_mat}
\title{Read in the counts}
\usage{
read_count_mat(fn)
}
\arguments{
\item{fn}{path to a file name typically a csv file.}
}
\value{
A matrix of counts of integers.
}
\description{
Read in the counts
}
\examples{
\dontrun{
library(here)
fn <- file.path("tests", "test_data", "gene_count_matrix_fixed.csv")
count_mat <- read_count_mat(fn)
}
}
