% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_template_parts.R
\name{template_yaml}
\alias{template_yaml}
\title{Create opening YAML for a report}
\usage{
template_yaml(
  title = "RNA-Seq Report",
  author = "Sameet",
  email = "sameet.mehta@yale.edu"
)
}
\arguments{
\item{title}{Title for the report}

\item{author}{Author for the report}

\item{email}{Email of the author for the report}
}
\value{
opening_yaml string to be used to construct the report qmd
}
\description{
Create opening YAML for a report
}
\examples{
\dontrun{
opnening_yaml <- template_yaml()
}
}
